Rest Client configuration
