+++ restClient/courses.http
@@ -23,4 +23,4 @@ Content-Type: application/json
 
 ###
 
-GET http://localhost:3002/coursesOfTeacher/4
+GET http://localhost:3002/teachers/4/courses
