+++ restClient/teachers.http
@@ -9,3 +9,7 @@ Content-Type: application/json
     "bio": "Outra Bio da Raisa Latorraca",
     "user": 4
     }
+
+###
+
+GET http://localhost:3002/teachers/4/courses
